requirements.txt
tk
pillow
pyinstaller